test-service/requirements.txt
flask>=3.1.1
ddtrace>=2.16.3
python-dotenv>=1.0.0

